Implement HomeWork04
Add file tests.py which have class "Test" that do looping of "check" test as many times as many numbers are present in the mapping dictionary for the chosen letter.
To run test use "python tests.py --letter a" command where "a" is letter that you want to choose from mapping dictionary.